0.4
Guardar y cargar json

+++ dataAccess/dao_file.tcl
@@ -1,28 +1,31 @@
-oo::class create daoFile {
-  # guardar list-br
-  # ip tuntap son volatiles
-  method saveTopology {fileName save} {
-    exec rm -rf save/$fileName
-    exec mkdir save/$fileName
-    set data [open save/$fileName/save.ncovs w]
-    puts $data $save
-    close $data
-  }
-
-  # cargamos 
-  # sudo ovs-vsctl get interface uml1_2 status si no es tun no se carga
-  method loadTopology {fileName} {
-      set fp [open $fileName r]
-      set file_data [read $fp]
-      close $fp
-      return $file_data
-  }
-  
-  method saveFlows {data fileName} {
-    if {[catch {exec echo $data > $fileName.ncovs} errmsg]} {
-      puts "$errmsg"
-      return 0
+if {[info exist daoFileExist] == 0} {
+  set daoFileExist 0
+  oo::class create daoFile {
+    # guardar list-br
+    # ip tuntap son volatiles
+    method saveTopology {fileName save} {
+      exec rm -rf save/$fileName
+      exec mkdir save/$fileName
+      set data [open save/$fileName/save.ncovs w]
+      puts $data $save
+      close $data
     }
-    return 1
+
+    # cargamos 
+    # sudo ovs-vsctl get interface uml1_2 status si no es tun no se carga
+    method loadTopology {fileName} {
+	set fp [open $fileName r]
+	set file_data [read $fp]
+	close $fp
+	return $file_data
+    }
+    
+    method saveFlows {data fileName} {
+      if {[catch {exec echo $data > $fileName.ncovs} errmsg]} {
+	puts "$errmsg"
+	return 0
+      }
+      return 1
+    }
   }
 }
